adding onto weapon framework. need to solve entity removal bugs

+++ SerenityLib/Engine/src/core/engine/renderer/opengl/State.cpp
@@ -14,7 +14,7 @@ OpenGLState::OpenGLState(const unsigned int& windowWidth, const unsigned int& wi
 OpenGLState::~OpenGLState() {
 }
 const unsigned int OpenGLState::getMaxTextureUnits() const {
-    return textureUnits.capacity();
+    return static_cast<unsigned int>(textureUnits.capacity());
 }
 void OpenGLState::GL_INIT_DEFAULT_STATE_MACHINE(const unsigned int& windowWidth, const unsigned int& windowHeight) {
     GLint     int_value;
